+++ packages/cli/src/commands/org/config/preview.ts
@@ -1,6 +1,8 @@
-import {DryRunOptions} from '@coveord/platform-client';
 import {Command, flags} from '@oclif/command';
+import {blueBright} from 'chalk';
+import {cli} from 'cli-ux';
 import {cwd} from 'process';
+import dedent from 'ts-dedent';
 import {
   buildAnalyticsFailureHook,
   buildAnalyticsSuccessHook,
@@ -10,11 +12,14 @@ import {
   IsAuthenticated,
   Preconditions,
 } from '../../../lib/decorators/preconditions';
+import {IsGitInstalled} from '../../../lib/decorators/preconditions/git';
+import {SnapshotOperationTimeoutError} from '../../../lib/errors';
 import {Snapshot} from '../../../lib/snapshot/snapshot';
 import {
   displayInvalidSnapshotError,
   displaySnapshotSynchronizationWarning,
   dryRun,
+  DryRunOptions,
   getTargetOrg,
   handleSnapshotError,
 } from '../../../lib/snapshot/snapshotCommon';
@@ -32,20 +37,27 @@ export default class Preview extends Command {
     }),
     showMissingResources: flags.boolean({
       char: 'd',
-      description: 'Whether or not preview missing resources',
+      description: 'Preview resources deletion when enabled',
       default: false,
+      required: false,
+    }),
+    snapshotId: flags.string({
+      char: 's',
+      description:
+        'The unique identifier of the snapshot to preview. If not specified, a new snapshot will be created from your local project. You can list available snapshots in your organization with org:config:list',
       required: false,
     }),
   };
 
   public static hidden = true;
 
-  @Preconditions(IsAuthenticated())
+  @Preconditions(IsAuthenticated(), IsGitInstalled())
   public async run() {
     const {flags} = this.parse(Preview);
     const target = await getTargetOrg(this.configuration, flags.target);
     const options: DryRunOptions = {
       deleteMissingResources: flags.showMissingResources,
+      snapshotId: flags.snapshotId,
     };
     const {reporter, snapshot, project} = await dryRun(
       target,
@@ -53,7 +65,7 @@ export default class Preview extends Command {
       options
     );
 
-    await snapshot.preview();
+    await snapshot.preview(project, options.deleteMissingResources);
 
     if (reporter.isSuccessReport()) {
       await snapshot.delete();
@@ -69,14 +81,31 @@ export default class Preview extends Command {
   public async catch(err?: Error) {
     const {flags} = this.parse(Preview);
     handleSnapshotError(err);
+    await this.displayAdditionalErrorMessage(err);
     await this.config.runHook(
       'analytics',
       buildAnalyticsFailureHook(this, flags, err)
     );
   }
 
+  private async displayAdditionalErrorMessage(err?: Error) {
+    if (err instanceof SnapshotOperationTimeoutError) {
+      const {flags} = this.parse(Preview);
+      const snapshot = err.snapshot;
+      const target = await getTargetOrg(this.configuration, flags.target);
+      cli.log(
+        dedent`
+
+          Once the snapshot is created, you can preview it with the following command:
+
+            ${blueBright`coveo org:config:preview -t ${target} -s ${snapshot.id}`}
+
+            `
+      );
+    }
+  }
+
   private async handleReportWithErrors(snapshot: Snapshot) {
-    // TODO: CDX-362: handle invalid snapshot cases
     const cfg = await this.configuration.get();
 
     if (snapshot.requiresSynchronization()) {