+++ packages/cli/src/lib/project/project.ts
@@ -1,34 +1,67 @@
-import {createWriteStream, existsSync, unlinkSync, writeFileSync} from 'fs';
-import {join} from 'path';
+import {
+  createWriteStream,
+  existsSync,
+  readdirSync,
+  unlinkSync,
+  writeFileSync,
+} from 'fs';
+import {extname, join} from 'path';
 import {cli} from 'cli-ux';
 import archiver from 'archiver';
 import {InvalidProjectError} from '../errors';
 import extract from 'extract-zip';
+import {DotFolder, DotFolderConfig} from './dotFolder';
+import {readJsonSync, writeJsonSync, WriteOptions} from 'fs-extra';
 
 export class Project {
-  public constructor(private pathToProject: string) {}
+  private static readonly resourceFolderName = 'resources';
+  public static readonly jsonFormat: WriteOptions = {spaces: '\t'};
+  public constructor(private _pathToProject: string) {
+    if (!this.isCoveoProject) {
+      this.makeCoveoProject();
+    }
+  }
 
   public async refresh(projectContent: Blob) {
     const buffer = await projectContent.arrayBuffer();
     const view = new DataView(buffer);
-    writeFileSync(this.pathToTemporaryZip, view);
-    await extract(this.pathToTemporaryZip, {dir: this.resourcePath});
+    writeFileSync(this.temporaryZipPath, view);
+    await extract(this.temporaryZipPath, {dir: this.resourcePath});
+    this.formatResourceFiles();
     this.deleteTemporaryZipFile();
   }
 
+  private formatResourceFiles(dirPath = this.resourcePath) {
+    const files = readdirSync(dirPath, {withFileTypes: true});
+    files.forEach((file) => {
+      const filePath = join(dirPath, file.name);
+      if (file.isDirectory()) {
+        this.formatResourceFiles(filePath);
+        return;
+      }
+      if (file.isFile() && extname(filePath) === '.json') {
+        const content = readJsonSync(filePath);
+        writeJsonSync(filePath, content, Project.jsonFormat);
+      }
+    });
+  }
+
   public deleteTemporaryZipFile() {
-    unlinkSync(this.pathToTemporaryZip);
+    unlinkSync(this.temporaryZipPath);
   }
 
   private ensureProjectCompliance() {
-    /*
-     * TODO: CDX-354: add checks to ensure the project is indeed a valid project
-     * e.g. * Check if path to resources is a folder
-     *      * Check if the root has a valid config file
-     */
-
-    if (!existsSync(this.resourcePath)) {
-      throw new InvalidProjectError();
+    if (!this.isResourcesProject) {
+      throw new InvalidProjectError(
+        this._pathToProject,
+        'Does not contain any resources folder'
+      );
+    }
+    if (!this.isCoveoProject) {
+      throw new InvalidProjectError(
+        this._pathToProject,
+        'Does not contain any .coveo folder'
+      );
     }
   }
 
@@ -36,8 +69,7 @@ export class Project {
     try {
       this.ensureProjectCompliance();
       await new Promise<void>((resolve, reject) => {
-        const pathToTemporaryZip = this.pathToTemporaryZip;
-        const outputStream = createWriteStream(pathToTemporaryZip);
+        const outputStream = createWriteStream(this.temporaryZipPath);
         const archive = archiver('zip');
 
         outputStream.on('close', () => resolve());
@@ -47,17 +79,37 @@ export class Project {
         archive.directory(this.resourcePath, false);
         archive.finalize();
       });
-      return this.pathToTemporaryZip;
+      return this.temporaryZipPath;
     } catch (error) {
       cli.error(error);
     }
   }
 
-  private get pathToTemporaryZip() {
-    return join(this.pathToProject, 'snapshot.zip');
+  public get pathToProject() {
+    return this._pathToProject;
   }
 
-  private get resourcePath() {
-    return join(this.pathToProject, 'resources');
+  private get temporaryZipPath() {
+    return join(this._pathToProject, 'snapshot.zip');
+  }
+
+  public get resourcePath() {
+    return join(this._pathToProject, Project.resourceFolderName);
+  }
+  public contains(fileName: string) {
+    return existsSync(join(this.pathToProject, fileName));
+  }
+
+  private get isCoveoProject() {
+    return this.contains(DotFolder.hiddenFolderName);
+  }
+
+  private get isResourcesProject() {
+    return this.contains(Project.resourceFolderName);
+  }
+
+  private makeCoveoProject() {
+    const dotFolder = new DotFolder(this.pathToProject);
+    new DotFolderConfig(dotFolder);
   }
 }